reexport moved names to simplify diff
