+++ ddtrace/internal/schema.py
@@ -1,0 +1,1 @@
+from ddtrace.internal.utils.schema import *  # noqa
