+++ ddtrace/internal/utils/wrapping/__init__.py
@@ -4,43 +4,27 @@ from types import FunctionType
 from typing import Any
 from typing import Callable
 from typing import Iterator
-from typing import MutableMapping
 from typing import Optional
 from typing import Protocol
 from typing import cast
-import weakref
 
 import bytecode as bc
 from bytecode import Instr
 
 from ddtrace.internal.assembly import Assembly
-from ddtrace.internal.threads import Lock
-from ddtrace.internal.wrapping.asyncs import wrap_async
-from ddtrace.internal.wrapping.generators import wrap_generator
+from ddtrace.internal.utils.inspection import link_function_to_code
+from ddtrace.internal.utils.wrapping.asyncs import wrap_async
+from ddtrace.internal.utils.wrapping.generators import wrap_generator
 
 
 PY = sys.version_info[:2]
 
-# Maps each wrapped function to its inner copy (the singly-linked list of
-# wrapping layers). WeakKeyDictionary so functions are not kept alive by the
-# registry alone.
-_wrapped: weakref.WeakKeyDictionary[FunctionType, FunctionType] = weakref.WeakKeyDictionary()
-_wrapped_lock = Lock()
-
-# Maps original code objects to the functions that own them. Written by
-# link_function_to_code; read by functions_for_code in inspection.py.
-# WeakValueDictionary so that wrapped ephemeral functions are not kept alive by
-# this mapping alone — inspection.py falls back to gc.get_referrers on a miss.
-_code_to_fn: MutableMapping[CodeType, FunctionType] = weakref.WeakValueDictionary()
-
-
-def link_function_to_code(code: CodeType, function: FunctionType) -> None:
-    """Link a function to its original code object for fast reverse lookup."""
-    _code_to_fn[code] = function
-
 
 class WrappedFunction(Protocol):
     """A wrapped function."""
+
+    __dd_wrapped__: Optional[FunctionType] = None
+    __dd_wrappers__: Optional[dict[Any, Any]] = None
 
     def __call__(self, *args: Any, **kwargs: Any) -> Any:
         pass
@@ -311,12 +295,11 @@ def wrap(f: FunctionType, wrapper: Wrapper) -> WrappedFunction:
         f.__defaults__,
         f.__closure__,
     )
-
-    # Carry forward the existing wrapped-function link to the new inner copy.
-    with _wrapped_lock:
-        existing_inner = _wrapped.get(f)
-        if existing_inner is not None:
-            _wrapped[wrapped] = existing_inner
+    try:
+        wf = cast(WrappedFunction, f)
+        cast(WrappedFunction, wrapped).__dd_wrapped__ = cast(FunctionType, wf.__dd_wrapped__)
+    except AttributeError:
+        pass
 
     wrapped.__kwdefaults__ = f.__kwdefaults__
 
@@ -346,54 +329,47 @@ def wrap(f: FunctionType, wrapper: Wrapper) -> WrappedFunction:
     # Replace the function code with the trampoline bytecode
     f.__code__ = wrapped_code.to_code()
 
-    # DEV: Multiple wrapping is implemented as a singly-linked list via _wrapped.
-    with _wrapped_lock:
-        _wrapped[f] = wrapped
+    # DEV: Multiple wrapping is implemented as a singly-linked list via the
+    # __dd_wrapped__ attribute.
+    wf = cast(WrappedFunction, f)
+    wf.__dd_wrapped__ = wrapped
 
     # Link the original code object to the original function
     link_function_to_code(code, f)
 
-    return cast(WrappedFunction, f)
-
-
-def _unwrap_method(f: Any) -> Any:
-    """Return the underlying function if *f* is a bound or unbound method."""
-    func = getattr(f, "__func__", f)
-    return func
+    return wf
 
 
 def is_wrapped(f: FunctionType) -> bool:
     """Check if a function is wrapped with any wrapper."""
     try:
-        with _wrapped_lock:
-            inner = _wrapped.get(_unwrap_method(f))
-    except TypeError:
-        # f is not weakly referenceable (e.g. C method_descriptor)
+        wf = cast(WrappedFunction, f)
+        inner = cast(FunctionType, wf.__dd_wrapped__)
+
+        # Sanity check
+        assert inner.__name__ == "<wrapped>", "Wrapper has wrapped function"  # nosec
+        return True
+    except AttributeError:
         return False
-    if inner is None:
-        return False
-    assert inner.__name__ == "<wrapped>", "Wrapper has wrapped function"  # nosec
-    return True
 
 
 def is_wrapped_with(f: FunctionType, wrapper: Wrapper) -> bool:
     """Check if a function is wrapped with a specific wrapper."""
-    f = cast(FunctionType, _unwrap_method(f))
     try:
-        with _wrapped_lock:
-            inner = _wrapped.get(f)
-    except TypeError:
+        wf = cast(WrappedFunction, f)
+        inner = cast(FunctionType, wf.__dd_wrapped__)
+
+        # Sanity check
+        assert inner.__name__ == "<wrapped>", "Wrapper has wrapped function"  # nosec
+
+        if wrapper in f.__code__.co_consts:
+            return True
+
+        # This is not the correct wrapping layer. Try with the next one.
+        return is_wrapped_with(inner, wrapper)
+
+    except AttributeError:
         return False
-    if inner is None:
-        return False
-
-    assert inner.__name__ == "<wrapped>", "Wrapper has wrapped function"  # nosec
-
-    if wrapper in f.__code__.co_consts:
-        return True
-
-    # This is not the correct wrapping layer. Try with the next one.
-    return is_wrapped_with(inner, wrapper)
 
 
 def unwrap(wf: WrappedFunction, wrapper: Wrapper) -> FunctionType:
@@ -403,52 +379,41 @@ def unwrap(wf: WrappedFunction, wrapper: Wrapper) -> FunctionType:
     this will unwrap the one that uses ``wrapper``. If the function was not
     wrapped with ``wrapper``, it will return the first argument.
     """
-    # DEV: Multiple wrapping layers are singly-linked via _wrapped. When we
-    # find the layer that needs to be removed we also have to ensure that we
+    # DEV: Multiple wrapping layers are singly-linked via __dd_wrapped__. When
+    # we find the layer that needs to be removed we also have to ensure that we
     # update the link at the deletion site if there is a non-empty tail.
-    f = cast(FunctionType, wf)
-    with _wrapped_lock:
-        inner = _wrapped.get(f)
-    if inner is None:
+    try:
+        inner = cast(FunctionType, wf.__dd_wrapped__)
+    except AttributeError:
         # The function is not wrapped so we return it as is.
-        return f
+        return cast(FunctionType, wf)
 
+    # Sanity check
     assert inner.__name__ == "<wrapped>", "Wrapper has wrapped function"  # nosec
 
-    if wrapper not in f.__code__.co_consts:
+    if wrapper not in cast(FunctionType, wf).__code__.co_consts:
         # This is not the correct wrapping layer. Try with the next one.
         return unwrap(cast(WrappedFunction, inner), wrapper)
 
-    # Remove the current wrapping layer by moving the next one over the current
-    # one. The code swap and registry update must be atomic: two concurrent
-    # unwrap calls on the same function and wrapper must not both succeed.
-    with _wrapped_lock:
-        f.__code__ = inner.__code__
-        next_inner = _wrapped.get(inner)
-        if next_inner is not None:
-            _wrapped[f] = next_inner
-        else:
-            del _wrapped[f]
+    # Remove the current wrapping layer by moving the next one over the
+    # current one.
+    f = cast(FunctionType, wf)
+    f.__code__ = inner.__code__
+
+    try:
+        # Update the link to the next layer.
+        wf.__dd_wrapped__ = cast(WrappedFunction, inner).__dd_wrapped__  # type: ignore[assignment]
+    except AttributeError:
+        # No more wrapping layers. Restore the original function by removing
+        # this extra attribute.
+        del wf.__dd_wrapped__
 
     return f
 
 
 def get_function_code(f: FunctionType) -> CodeType:
-    with _wrapped_lock:
-        inner = _wrapped.get(f)
-    return (inner if inner is not None else f).__code__
+    return (cast(WrappedFunction, f).__dd_wrapped__ or f if is_wrapped(f) else f).__code__
 
 
 def set_function_code(f: FunctionType, code: CodeType) -> None:
-    with _wrapped_lock:
-        inner = _wrapped.get(f)
-    (inner if inner is not None else f).__code__ = code
-
-
-def get_wrapped(f: FunctionType) -> Optional[FunctionType]:
-    """Return the inner bytecode copy of *f* if it is wrapped, else None."""
-    try:
-        with _wrapped_lock:
-            return _wrapped.get(cast(FunctionType, _unwrap_method(f)))
-    except TypeError:
-        return None
+    (cast(WrappedFunction, f).__dd_wrapped__ or f if is_wrapped(f) else f).__code__ = code  # type: ignore[misc]
